pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sparta-pipeline"
dynamic = ["version"]
authors = [{name = "Baptiste Ruiz", email = "baptiste.ruiz@inria.fr"}]
readme = "README.md"
description = "A pipeline for Machine-Learning based analysis of functional and taxonomic profilings of microbial communities"
license = {text = "GNU General Public License v3.0"}

dependencies = [
  'pandas',
  'numpy',
  'scikit-learn',
  'scipy',
  'matplotlib',
  'seaborn',
  'joblib',
  'tqdm',
  'goatools',
  'biopython',
  'requests',
  'kneebow',
  'esmecata',
  'shap'
]

[project.scripts]
sparta = "sparta_pipeline.__main__:main"

[project.urls]
Homepage = "https://github.com/baptisteruiz/SPARTA"

[tool.setuptools]
packages = ['sparta_pipeline']
package-dir = {'sparta_pipeline' = 'sparta_pipeline'}

[tool.setuptools.dynamic]
version = { attr = "sparta_pipeline.__version__" }
dependencies = {file = ["requirements.txt"]}